schwarz-precond: merge buffers into executor
The buffer-pooling types (BufferPool, LocalSolveScratch,
AdditiveSweepBuffers, SchwarzBuffers, WorkerReductionBuffers) were
used only by AdditiveExecutor, so the file boundary served no
isolation purpose. Fold them into executor.rs and tighten visibility:
all buffer types and helpers become file-private (no pub(super)),
since they're only touched by AdditiveExecutor. Only AdditiveExecutor
itself plus the seven fields/methods preconditioner.rs actually
accesses retain pub(super).
